Apply improved types & architecture to Crypto_org
