+++ libs/ledger-live-common/src/families/crypto_org/broadcast.ts
@@ -1,7 +1,8 @@
-import type { Operation, SignedOperation, Account } from "@ledgerhq/types-live";
+import type { AccountBridge } from "@ledgerhq/types-live";
 import { patchOperationWithHash } from "../../operation";
 import { broadcastTransaction } from "./api";
 import { CryptoOrgErrorBroadcasting } from "./errors";
+import { Transaction } from "./types";
 
 function isBroadcastTxFailure(result) {
   return !!result.code;
@@ -10,13 +11,10 @@ function isBroadcastTxFailure(result) {
 /**
  * Broadcast the signed transaction
  */
-const broadcast = async ({
+const broadcast: AccountBridge<Transaction>["broadcast"] = async ({
   account,
   signedOperation: { signature, operation },
-}: {
-  account: Account;
-  signedOperation: SignedOperation;
-}): Promise<Operation> => {
+}) => {
   const broadcastResponse = await broadcastTransaction(signature, account.currency.id);
 
   if (isBroadcastTxFailure(broadcastResponse)) {
