+++ libs/ledger-live-common/src/families/crypto_org/getTransactionStatus.ts
@@ -1,0 +1,65 @@
+import { BigNumber } from "bignumber.js";
+import {
+  NotEnoughBalance,
+  RecipientRequired,
+  InvalidAddress,
+  FeeNotLoaded,
+  AmountRequired,
+  InvalidAddressBecauseDestinationIsAlsoSource,
+} from "@ledgerhq/errors";
+import type { AccountBridge } from "@ledgerhq/types-live";
+import type { Transaction } from "./types";
+import { isValidAddress } from "./logic";
+
+const getTransactionStatus: AccountBridge<Transaction>["getTransactionStatus"] = async (
+  account,
+  transaction,
+) => {
+  const errors: {
+    fees?: Error;
+    amount?: Error;
+    recipient?: Error;
+  } = {};
+  const warnings = {};
+  const useAllAmount = !!transaction.useAllAmount;
+
+  if (!transaction.fees) {
+    errors.fees = new FeeNotLoaded();
+  }
+
+  const estimatedFees = transaction.fees || new BigNumber(0);
+  const totalSpent = useAllAmount
+    ? account.balance
+    : new BigNumber(transaction.amount).plus(estimatedFees);
+  const amount = useAllAmount
+    ? account.balance.minus(estimatedFees)
+    : new BigNumber(transaction.amount);
+
+  if (totalSpent.gt(account.balance)) {
+    errors.amount = new NotEnoughBalance();
+  }
+
+  if (!amount.gt(0)) {
+    errors.amount = new AmountRequired();
+  }
+
+  if (!transaction.recipient) {
+    errors.recipient = new RecipientRequired();
+  } else if (!isValidAddress(transaction.recipient, account.currency.id)) {
+    errors.recipient = new InvalidAddress("", {
+      currencyName: account.currency.name,
+    });
+  } else if (transaction.mode === "send" && account.freshAddress === transaction.recipient) {
+    errors.recipient = new InvalidAddressBecauseDestinationIsAlsoSource();
+  }
+
+  return Promise.resolve({
+    errors,
+    warnings,
+    estimatedFees,
+    amount,
+    totalSpent,
+  });
+};
+
+export default getTransactionStatus;
